Moved View into components, deleted Loader.js
